requirements.txt
# Core FastAPI and web framework dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Database dependencies
mysql-connector-python>=8.2.0
pymysql>=1.1.0

# OpenTelemetry dependencies
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-mysql>=0.42b0
opentelemetry-exporter-otlp-proto-http>=1.21.0

# Additional useful libraries
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Development and testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
black>=23.0.0
flake8>=6.0.0 